Добавлено:     * Рендер теперь использует буфер глубины для Vulkan.
Исправлено:
    * Работа функции получения строки события по его коду.
    * Способ задания размеров окна при создании для Windows.
    * По кнопке закрытия окна обрабатывается запрос закрытия для Windows.
    * Изменен способ перестройки цепочки обмена для Vulkan.
    * Изменен список предпочитаемых форматов буфера глубины.

+++ engine/src/core/event.c
@@ -196,7 +196,7 @@ API const char* event_code_to_str(event_code code)
 
     static const u32 code_count = sizeof(strings) / sizeof(char*);
 
-    if(code > code_count || strings[code] == nullptr)
+    if((u32)code > code_count || strings[code] == nullptr)
     {
         return "UNKNOWN";
     }
